feat: The added View present a Lazy Horizontal ScrollView
This adds the functionality for the user to swipe through the scroll view horizontally with pinned views at the start of the Hstack and end of the Hstack which will not present until all images are scrolled through. Meaning, the pin that is leading will stay until the ending pin is shown. This could be useful for presenting long, lazy scroll views and giving more context as the user scrolls